backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Data validation and config
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File processing
python-magic==0.4.27
Pillow>=10.0.0
imagehash>=4.3.1
exifread>=3.1.2

# Machine learning for similarity analysis
numpy>=1.24.0
scikit-learn>=1.3.0

# Web and utilities
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
requests>=2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
